[FLINK-14942][state-processor-api] move to output package
