+++ flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/WritableSavepoint.java
@@ -23,8 +23,8 @@ import org.apache.flink.api.java.io.DiscardingOutputFormat;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.runtime.checkpoint.OperatorState;
 import org.apache.flink.runtime.state.StateBackend;
-import org.apache.flink.state.api.functions.FileCopyMapFunction;
-import org.apache.flink.state.api.functions.StatePathExtractor;
+import org.apache.flink.state.api.output.FileCopyMapFunction;
+import org.apache.flink.state.api.output.StatePathExtractor;
 import org.apache.flink.state.api.output.MergeOperatorStates;
 import org.apache.flink.state.api.output.SavepointOutputFormat;
 import org.apache.flink.state.api.runtime.BootstrapTransformationWithID;
