+++ flink-libraries/flink-state-processing-api/src/main/java/org/apache/flink/state/api/output/FileCopyMapFunction.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.state.api.functions;
+package org.apache.flink.state.api.output;
 
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.core.fs.Path;
